backend/requirements.txt
pyswisseph==2.10.3.2
pytz==2023.3
geopy==2.4.1
flask==2.3.3
flask-cors==4.0.0
timezonefinder==6.5.9
requests==2.31.0
openai==1.14.3
python-dotenv==1.0.1
